Error of Reduced-Rank Canonical Variate Analysis

cva_error calculates the errors of a canonical-variate regression model built on a training set and applied to a test set.

cva_error(x, y, x_new, y_new, rank = "full", k = 0)

Arguments

x

data frame or matrix of predictor variables

y

data frame or matrix of response variables

x_new

data frame or matrix of input variables used to predict the response matrix.

y_new

data frame or matrix of observed response variables.

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

Value

data frame of error.

References

Izenman, A.J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

set.seed(12345)
library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_train <-
sample_size <- floor(0.66 * nrow(galaxy))
train_ind <- sample(seq_len(nrow(galaxy)), size = sample_size)
train <- galaxy[train_ind, ]
test <- galaxy[-train_ind, ]
train_x <- select(train, -Rmag:-chi2red)
train_y <- select(train, Rmag:chi2red)
test_x <- select(test, -Rmag:-chi2red)
test_y <- select(test, Rmag:chi2red)
cva_error(train_x, train_y, test_x, test_y, rank = 2, k = 0.001)

#> # A tibble: 1,178 × 3
#>    index        CV1         CV2
#>    <int>      <dbl>       <dbl>
#> 1      1 -0.3325655  0.56296266
#> 2      2  0.2136488  0.21247374
#> 3      3 -0.7341411  1.46165807
#> 4      4  0.2707906  0.59126916
#> 5      5 -0.1121691  0.64194445
#> 6      6 -0.6899002 -0.19232415
#> 7      7  0.1226567 -0.21403048
#> 8      8  0.7249506  0.03342818
#> 9      9  0.3965837  0.81853987
#> 10    10 -0.5252940  1.56516986
#> # ... with 1,168 more rows