Canonical Variate Scores

cva_scores creates linear combinations of predictor and response variables from the coefficients of reduced-rank canonical variate analysis.

cva_scores(x, y, rank = "full", k = 0)

Arguments

x

data frame or matrix of predictor variables

y

data frame or matrix of response variables

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

Value

list containing: data frame of canonical variate scores of predictor variables; data frame of canonical variate scores of response variables; vector of canonical correlations.

References

Izenman, A.J. (2008) Modern Multivariate Statistical Techniques. Springer.

See also

cva

Examples

library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cva_scores(galaxy_x, galaxy_y, rank = 2)

#> $xi
#> # A tibble: 3,462 × 2
#>            xi1        xi2
#>          <dbl>      <dbl>
#> 1  -0.08238313 -0.6676924
#> 2   0.02260035 -1.4511158
#> 3   1.11408330 -0.2493058
#> 4  -0.05463705  0.2490659
#> 5  -0.13694910 -0.2669298
#> 6   0.48773287 -0.9107072
#> 7  -0.42326056 -0.5505199
#> 8  -0.56698136 -0.3770152
#> 9  -0.08239422 -0.2889443
#> 10 -2.98040950  1.1501098
#> # ... with 3,452 more rows
#> 
#> $omega
#> # A tibble: 3,462 × 2
#>        omega1     omega2
#>         <dbl>      <dbl>
#> 1   0.1961662 -0.7469899
#> 2   0.3706709 -1.1484080
#> 3   0.6920836 -0.2842400
#> 4  -0.2791397 -0.9053483
#> 5   0.6305252 -1.2606469
#> 6   1.6221327 -0.6090875
#> 7  -0.7805154 -1.0548006
#> 8  -0.4741124 -0.8876047
#> 9  -0.3835437 -0.4417843
#> 10 -2.3198192  0.4277744
#> # ... with 3,452 more rows
#> 
#> $canonical_corr
#> [1] 0.94263015 0.57868061 0.08584486 0.04088308 0.03102955 0.02483753
#>