Fit Reduced-Rank LDA Model

lda fits a reduced-rank linear discriminant analysis model.

lda(x, class, rank = "full", type = "cov", k = 0, quadratic = FALSE)

Arguments

x

data frame or matrix of input variables

class

vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

type

the format of the covariance matrix. type = "cov" runs the regression using the mean-centered covariance matrix. type = "cor" runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

quadratic

`logical`. If `TRUE` fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.

Value

`list` containing: a data frame of class

References

Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(pendigits)
digits_features <- pendigits[,-35:-36]
digits_class <- pendigits[,35]
lda(digits_features, digits_class, rank = 3, k = 0.001)

#> $G
#> # A tibble: 34 × 3
#>                 LD1              LD2              LD3
#>              <cplx>           <cplx>           <cplx>
#> 1  -4.084331e-04+0i  2.673256e-03+0i -2.828805e-03+0i
#> 2   1.318003e-04+0i  1.744069e-03+0i -7.641535e-03+0i
#> 3   1.760460e-04+0i  2.251182e-04+0i -4.268579e-05+0i
#> 4   2.674138e-04+0i -4.999458e-03+0i  1.070153e-02+0i
#> 5  -6.090273e-05+0i -2.958977e-03+0i  6.886435e-03+0i
#> 6  -1.871425e-04+0i  1.660323e-03+0i -3.183527e-04+0i
#> 7   1.024797e-04+0i -1.769586e-03+0i -1.934749e-03+0i
#> 8   9.924649e-05+0i  3.954015e-05+0i  1.137345e-03+0i
#> 9  -3.667089e-05+0i  3.541013e-04+0i  3.179892e-03+0i
#> 10 -6.109829e-04+0i  7.364155e-04+0i  4.902319e-03+0i
#> # ... with 24 more rows
#> 
#> $H
#> # A tibble: 9 × 3
#>              LD1            LD2           LD3
#>           <cplx>         <cplx>        <cplx>
#> 1 -2.85634163+0i  1.51237682+0i  0.8050839+0i
#> 2 -2.50054364+0i -1.09099530+0i  0.5518140+0i
#> 3 -2.08040344+0i -1.75985448+0i  0.2875693+0i
#> 4 -1.84678718+0i -0.37414635+0i  1.1998905+0i
#> 5 -1.53932365+0i -0.08673702+0i -1.3482415+0i
#> 6 -1.13402336+0i  0.33147944+0i  1.0048436+0i
#> 7 -0.73595413+0i  0.41396783+0i -0.7817387+0i
#> 8 -0.33291438+0i -1.27968435+0i  0.9148828+0i
#> 9 -0.04292934+0i  0.39895558+0i  1.4905944+0i
#> 
#> $eigen_portion
#> [1] 0.3755987+0i 0.3394781+0i 0.2849233-0i
#>