Linear Discriminant Scores

lda_scores calculates canonical-variate scores. In the case of linear discriminant analysis, the canonical-variate scores of the predictor variables are the linear discriminant functions evaluated at the values of predictor variables, while the scores for the response variables correspond to class means.

lda_scores(x, class, rank = "full", type = "cov", k = 0,
  quadratic = FALSE)

Arguments

x

data frame or matrix of predictor variables

class

vector or one-column data frame of type factor or character that are the class labels of the observations.

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

type

the format of the covariance matrix. type = "cov" runs the regression using the mean-centered covariance matrix. type = "cor" runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

quadratic

logical. If TRUE fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.

Value

list containing: data frame of linear discriminant function scores; data frame of class means.

References

Izenman, A.J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(iris)
lda_scores(iris[,1:4], iris[,5])

#> $scores
#> # A tibble: 150 × 3
#>          LD1          LD2  class
#>        <dbl>        <dbl> <fctr>
#> 1  -1.387419  0.125704003 setosa
#> 2  -1.226832 -0.329159709 setosa
#> 3  -1.288983 -0.111043950 setosa
#> 4  -1.172537 -0.280609930 setosa
#> 5  -1.399553  0.215264848 setosa
#> 6  -1.325489  0.611623049 setosa
#> 7  -1.241276  0.148891363 setosa
#> 8  -1.308855 -0.004867908 setosa
#> 9  -1.129057 -0.424771543 setosa
#> 10 -1.263725 -0.396383630 setosa
#> # ... with 140 more rows
#> 
#> $class_means
#> # A tibble: 3 × 3
#>          LD1        LD2      class
#>        <dbl>      <dbl>     <fctr>
#> 1 -1.3499219  0.4054351     setosa
#> 2  0.3238438 -1.3717842 versicolor
#> 3  1.0260780  0.9663491  virginica
#>