pca

pca fits a reduced-rank principal component model. This is a special case of reduced-rank regression with the weight matrix set the identity, i.e., \(\mathbf{\Gamma} = \mathbf{I}\), and response variables set to the predictor variables, i.e, \(\mathbf{X} \equiv \mathbf{Y}\). Principal componennt analysis creates a new variable, the first principal component, that is a linear combination of the origianl predictor variables that maximizes variance. The second principal component is constructed as a linear combination of the original predictor variables while being uncorrelated with the first principal component. There are, at most, as many principal components as original predictor variables, each of which are uncorrelated with each i-th principal component capture less variance than the i - 1 principal components.

pca(x, rank = "full", k = 0)

Value

list containing: named vector of means; data frame of reduced-rank regression coefficients; data frame of principal components; named matrix of goodness-of-fit statistics for the principal components.