Fit Reduced-Rank Regression Model

rrr fits a reduced-rank regression model.

rrr(x, y, type = "identity", rank = "full", k = 0)

Arguments

x

data frame or matrix of predictor variables

y

data frame or matrix of response variables

type

type of reduced-rank regression model to fit. type = "identity", the default, uses \(\mathbf{\Gamma} = \mathbf{I}\) to fit a reduced-rank regression. type = "pca" fits a principal component analysis model as a special case of reduced-rank regression. type = "cva" fits a canonical variate analysis model as a special case of reduced-rank regression. type = "lda" fits a linear discriminant analysis model as a special case of reduced-rank regression.

rank

rank of coefficient matrix.

k

small constant added to diagonal of covariance matrices to make inversion easier.

Value

list containing estimates of coefficients and means, and eigenvalue-based diagnostics.

References

Izenman, A.J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(tobacco)
tobacco_x <- tobacco[,4:9]
tobacco_y <- tobacco[,1:3]
rrr(tobacco_x, tobacco_y, rank = 1)

#> $mean
#>                         [,1]
#> Y1.BurnRate         1.749756
#> Y2.PercentSugar    14.687990
#> Y3.PercentNicotine  2.639914
#> 
#> $A
#>                           [,1]
#> Y1.BurnRate         0.03107787
#> Y2.PercentSugar    -0.97005030
#> Y3.PercentNicotine  0.24090782
#> 
#> $B
#>      X1.PercentNitrogen X2.PercentChlorine X3.PercentPotassium
#> [1,]            4.32427          -1.358648           -1.480832
#>      X4.PercentPhosphorus X5.PercentCalcium X6.PercentMagnesium
#> [1,]            -13.72942        -0.4528289            3.866896
#> 
#> $C
#>                    X1.PercentNitrogen X2.PercentChlorine X3.PercentPotassium
#> Y1.BurnRate                 0.1343891        -0.04222389         -0.04602109
#> Y2.PercentSugar            -4.1947590         1.31795724          1.43648115
#> Y3.PercentNicotine          1.0417504        -0.32730901         -0.35674392
#>                    X4.PercentPhosphorus X5.PercentCalcium X6.PercentMagnesium
#> Y1.BurnRate                  -0.4266812       -0.01407296           0.1201749
#> Y2.PercentSugar              13.3182319        0.43926680          -3.7510833
#> Y3.PercentNicotine           -3.3075256       -0.10909002           0.9315654
#> 
#> $eigenvalues
#> [1] 3.28209974 0.03782978 0.01015996
#> 

data(pendigits)
digits_features <- pendigits[, -35:-36]
rrr(digits_features, digits_features, type = "pca", rank = 3)

#> $means
#>        V1        V2        V3        V4        V5        V6        V7        V8 
#> 38.814320 85.120269 40.605622 83.774199 49.770378 65.573144 51.220251 44.498999 
#>        V9       V10       V11       V12       V13       V14       V15       V16 
#> 56.868541 33.695961 60.516376 34.826510 55.022289 34.937045 47.287482 28.845342 
#>       V17       V18       V19       V20       V21       V22       V23       V24 
#>  4.431587  5.341794 38.814320 85.120269 40.605622 83.774199 49.770378 65.573144 
#>       V25       V26       V27       V28       V29       V30       V31       V32 
#> 51.220251 44.498999 56.868541 33.695961 60.516376 34.826510 55.022289 34.937045 
#>       V33       V34 
#> 47.287482 28.845342 
#> 
#> $C
#> # A tibble: 34 × 34
#>             V1           V2           V3           V4           V5           V6
#>          <dbl>        <dbl>        <dbl>        <dbl>        <dbl>        <dbl>
#> 1   0.15227028  0.041971586  0.057621563  0.032824603 -0.083070856 -0.005097600
#> 2   0.04197159  0.013118640  0.010598715  0.005435764 -0.027861499 -0.010118470
#> 3   0.05762156  0.010598715  0.057059768  0.042972135  0.005136858  0.048612391
#> 4   0.03282460  0.005435764  0.042972135  0.034784478  0.014445469  0.041246125
#> 5  -0.08307086 -0.027861499  0.005136858  0.014445469  0.083611542  0.054433598
#> 6  -0.00509760 -0.010118470  0.048612391  0.041246125  0.054433598  0.074336749
#> 7  -0.12925579 -0.039313730 -0.047145641 -0.030698476  0.070011543  0.011999221
#> 8  -0.03613352 -0.021185046  0.025475341  0.019298193  0.056664934  0.065381821
#> 9  -0.04375152 -0.018001465 -0.036555354 -0.038171115 -0.002983700 -0.013777198
#> 10 -0.05163631 -0.020586646 -0.022911782 -0.022807068  0.019987392  0.007198665
#> # ... with 24 more rows, and 28 more variables: V7 <dbl>, V8 <dbl>, V9 <dbl>,
#> #   V10 <dbl>, V11 <dbl>, V12 <dbl>, V13 <dbl>, V14 <dbl>, V15 <dbl>,
#> #   V16 <dbl>, V17 <dbl>, V18 <dbl>, V19 <dbl>, V20 <dbl>, V21 <dbl>,
#> #   V22 <dbl>, V23 <dbl>, V24 <dbl>, V25 <dbl>, V26 <dbl>, V27 <dbl>,
#> #   V28 <dbl>, V29 <dbl>, V30 <dbl>, V31 <dbl>, V32 <dbl>, V33 <dbl>, V34 <dbl>
#> 
#> $PC
#> # A tibble: 34 × 3
#>            PC1          PC2         PC3
#>          <dbl>        <dbl>       <dbl>
#> 1   0.02861158 -0.142768395 -0.36203431
#> 2   0.04680878 -0.033512929 -0.09901748
#> 3  -0.13901701  0.046687918 -0.18855842
#> 4  -0.10267046  0.083217740 -0.13159811
#> 5  -0.15980723  0.196663538  0.13927183
#> 6  -0.24050660  0.117714673 -0.05134764
#> 7  -0.10613246  0.031714447  0.33613210
#> 8  -0.28960021 -0.002314316  0.07783244
#> 9  -0.11746249 -0.263105908  0.21532196
#> 10 -0.14441246 -0.150734547  0.19065755
#> # ... with 24 more rows
#> 
#> $goodness_of_fit
#>       PC1       PC2       PC3 
#> 0.7168777 0.4680599 0.3144671 
#> 

library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
rrr(galaxy_x, galaxy_y, type = "cva", rank = 2)

#> $mean
#>               [,1]
#> Rmag    27.0430945
#> ApDRmag  0.9028657
#> mumax   25.6729200
#> Mcz     -1.3514412
#> MCzml   -1.2214416
#> chi2red  2.1352826
#> 
#> $G
#>           UjMAG        BjMAG     VjMAG       usMAG      gsMAG      rsMAG
#> [1,] -0.4026736 -0.002648682 -1.063946 -0.03793168  0.1435434  0.7446434
#> [2,] -1.8871230 -0.021421754  3.036531 -1.92938162 -0.8238762 -2.4591042
#>           UbMAG       BbMAG       VnMAG      S280MAG    W420FE    W462FE
#> [1,] 0.03191507  0.13435547 -0.01074357 -0.005520653  1.808397  2.594173
#> [2,] 4.20231673 -0.02961893  0.06238929 -0.031742088 32.171809 12.197256
#>         W485FD     W518FE    W571FS    W604FE    W646FD    W696FE     W753FE
#> [1,]  2.794240  -1.279663   1.52135 -1.927474 -1.306194 -1.297319  -3.754107
#> [2,] -4.094297 -18.010147 -16.43018 -2.484012  6.844791 15.661076 -11.090582
#>          W815FS    W856FD    W914FD     W914FE
#> [1,] -0.6842338 0.7212811  4.543466  0.1558235
#> [2,] 14.5603371 3.6135546 -6.306661 -2.3404624
#> 
#> $H
#>            Rmag    ApDRmag      mumax        Mcz      MCzml    chi2red
#> [1,] -0.5657905 -0.4822494  0.1134821  1.4069068  1.3645567 -0.6210710
#> [2,] -0.2869686  0.1150591 -0.4323321 -0.5932063 -0.5813512  0.2503455
#> 
#> $canonical_corr
#> [1] 0.94263015 0.57868061 0.08584486 0.04088308 0.03102955 0.02483753
#> 

data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
rrr(iris_x, iris_y, type = "lda")

#> $G
#> # A tibble: 4 × 2
#>          LD1         LD2
#>        <dbl>       <dbl>
#> 1 -0.1427341  0.01008498
#> 2 -0.2640795  0.90569343
#> 3  0.3788238 -0.38994070
#> 4  0.4836742  1.18799194
#> 
#> $H
#> # A tibble: 2 × 2
#>          LD1        LD2
#>        <dbl>      <dbl>
#> 1 -1.1238094 -0.3321454
#> 2 -0.2653032 -1.1058991
#>