Compute Latent Variable Scores

Compute Latent Variable Scores

scores(x, y, type = "pca", rank = "full", k = 0)

Arguments

x

data frame or matrix of predictor variables

y

data frame or matrix of response variables

type

type of reduced-rank regression model to fit. type = "identity", the default, uses \(\mathbf{\Gamma} = \mathbf{I}\) to fit a reduced-rank regression. type = "pca" fits a principal component analysis model as a special case of reduced-rank regression. type = "cva" fits a canonical variate analysis model as a special case of reduced-rank regression. type = "lda" fits a linear discriminant analysis model as a special case of reduced-rank regression.

rank

rank of coefficient matrix.

k

small constant added to diagonal of covariance matrices to make inversion easier.

References

Izenman, A.J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(pendigits)
digits_features <- pendigits[, -35:-36]
scores(digits_features, digits_features, type = "pca", rank = 3)

#> # A tibble: 10,992 × 3
#>            PC1         PC2        PC3
#>          <dbl>       <dbl>      <dbl>
#> 1   155.566449    4.978394  -29.92034
#> 2   -13.769794  107.245141  -29.49099
#> 3  -106.072574  -12.394709  110.20702
#> 4    -4.184733  -47.779421   11.43866
#> 5  -100.904599   50.618077   76.39842
#> 6   -19.235101   31.016531 -151.44893
#> 7    21.368904  -63.616851  122.34104
#> 8   122.842132 -128.821329  113.64392
#> 9   172.378851  111.913854   65.54326
#> 10  142.956443 -100.197126  -25.96261
#> # ... with 10,982 more rows

library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
scores(galaxy_x, galaxy_y, type = "cva", rank = 4)

#> $xi
#> # A tibble: 3,462 × 4
#>            xi1        xi2         xi3         xi4
#>          <dbl>      <dbl>       <dbl>       <dbl>
#> 1  -0.08238313 -0.6676924  0.02515131 -0.02481413
#> 2   0.02260035 -1.4511158 -0.58088584  0.20733938
#> 3   1.11408330 -0.2493058  0.09403073 -0.07837298
#> 4  -0.05463705  0.2490659 -0.20970644  0.01769805
#> 5  -0.13694910 -0.2669298  0.96961627  0.25516577
#> 6   0.48773287 -0.9107072  0.23605276  0.27926709
#> 7  -0.42326056 -0.5505199 -0.29023430 -0.01203045
#> 8  -0.56698136 -0.3770152  0.45828666  0.13158831
#> 9  -0.08239422 -0.2889443  0.04667063 -0.12539483
#> 10 -2.98040950  1.1501098 -0.18441986 -0.05890351
#> # ... with 3,452 more rows
#> 
#> $omega
#> # A tibble: 3,462 × 4
#>         omega1      omega2      omega3     omega4
#>          <dbl>       <dbl>       <dbl>      <dbl>
#> 1  -0.28752003 -0.41692284  1.87565068  0.7635578
#> 2   0.25036357 -1.07101396  0.99951862  1.7260456
#> 3   1.13153635 -0.59329585 -0.66460774  2.3022547
#> 4   0.01776948 -1.11898076  0.09758743  3.1308931
#> 5  -0.75972796 -0.29860131  3.87975711 -2.1613321
#> 6   0.79193245 -0.03965822  2.89060392  0.3630299
#> 7  -0.56121833 -1.20667935 -0.59762529  0.3823229
#> 8  -0.15715388 -1.09962079 -1.71365870 -1.8968769
#> 9  -0.26928852 -0.52238472 -0.14473095  0.6794540
#> 10 -2.48793585  0.53335612  1.68773178  3.2506884
#> # ... with 3,452 more rows
#> 
#> $canonical_corr
#> [1] 0.94263015 0.57868061 0.08584486 0.04088308 0.03102955 0.02483753
#> 

data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
scores(iris_x, iris_y, type = "lda")

#> $scores
#> # A tibble: 150 × 3
#>          LD1          LD2  class
#>        <dbl>        <dbl> <fctr>
#> 1  -1.387419  0.125704003 setosa
#> 2  -1.226832 -0.329159709 setosa
#> 3  -1.288983 -0.111043950 setosa
#> 4  -1.172537 -0.280609930 setosa
#> 5  -1.399553  0.215264848 setosa
#> 6  -1.325489  0.611623049 setosa
#> 7  -1.241276  0.148891363 setosa
#> 8  -1.308855 -0.004867908 setosa
#> 9  -1.129057 -0.424771543 setosa
#> 10 -1.263725 -0.396383630 setosa
#> # ... with 140 more rows
#> 
#> $class_means
#> # A tibble: 3 × 3
#>          LD1        LD2      class
#>        <dbl>      <dbl>     <fctr>
#> 1 -1.3499219  0.4054351     setosa
#> 2  0.3238438 -1.3717842 versicolor
#> 3  1.0260780  0.9663491  virginica
#>