Fit Reduced-Rank LDA Model

lda fits a reduced-rank linear discriminant analysis model.

lda(x, class, rank = "full", type = "cov", k = 0, quadratic = FALSE)

Arguments

x

data frame or matrix of input variables

class

vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

type

the format of the covariance matrix. type = "cov" runs the regression using the mean-centered covariance matrix. type = "cor" runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

quadratic

`logical`. If `TRUE` fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.

Value

`list` containing: a data frame of class

References

Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(pendigits)
digits_features <- pendigits[,-35:-36]
digits_class <- pendigits[,35]
lda(digits_features, digits_class, rank = 3, k = 0.001)

#> $prior
#>          0          1          2          3          4          5          6 
#> 0.10398472 0.10398472 0.10407569 0.09597889 0.10407569 0.09597889 0.09606987 
#>          7          8          9 
#> 0.10389374 0.09597889 0.09597889 
#> 
#> $G
#> # A tibble: 34 × 3
#>              LD1           LD2           LD3
#>            <dbl>         <dbl>         <dbl>
#> 1  -4.084331e-04  2.673256e-03 -2.828805e-03
#> 2   1.318003e-04  1.744069e-03 -7.641535e-03
#> 3   1.760460e-04  2.251182e-04 -4.268579e-05
#> 4   2.674138e-04 -4.999458e-03  1.070153e-02
#> 5  -6.090273e-05 -2.958977e-03  6.886435e-03
#> 6  -1.871425e-04  1.660323e-03 -3.183527e-04
#> 7   1.024797e-04 -1.769586e-03 -1.934749e-03
#> 8   9.924649e-05  3.954015e-05  1.137345e-03
#> 9  -3.667089e-05  3.541013e-04  3.179892e-03
#> 10 -6.109829e-04  7.364155e-04  4.902319e-03
#> # ... with 24 more rows
#> 
#> $H
#> # A tibble: 9 × 3
#>           LD1         LD2        LD3
#>         <dbl>       <dbl>      <dbl>
#> 1 -2.85634163  1.51237682  0.8050839
#> 2 -2.50054364 -1.09099530  0.5518140
#> 3 -2.08040344 -1.75985448  0.2875693
#> 4 -1.84678718 -0.37414635  1.1998905
#> 5 -1.53932365 -0.08673702 -1.3482415
#> 6 -1.13402336  0.33147944  1.0048436
#> 7 -0.73595413  0.41396783 -0.7817387
#> 8 -0.33291438 -1.27968435  0.9148828
#> 9 -0.04292934  0.39895558  1.4905944
#> 
#> $eigen_portion
#> [1] 0.3755987 0.3394781 0.2849233
#>