Fit Reduced-Rank LDA Model

lda fits a reduced-rank linear discriminant analysis model by fitting a reduced-rank canonical variate model to the predictor variables and a response-indicator matrix.

lda(x, class, rank = "full", k = 0)

Arguments

x

data frame or matrix of predictor variables

class

vector or one-column data frame of type factor or character that are the class labels of the observations.

rank

of the coefficient matrix to estimate. Default rank = full prodcues the standard multivariate regression technique.

k

small constant to add to the ridge of covariance matrices to avoid singularity.

Value

list containing: a named vector of prior probabilities; a data frame of canonical-variate coefficients of the predictor variables; a data frame of canonical-variate coefficients for the reponse variables.

References

Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.

Examples

data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
lda(iris_x, iris_y)

#> $G
#> # A tibble: 4 × 2
#>          LD1         LD2
#>        <dbl>       <dbl>
#> 1 -0.1427341  0.01008498
#> 2 -0.2640795  0.90569343
#> 3  0.3788238 -0.38994070
#> 4  0.4836742  1.18799194
#> 
#> $H
#> # A tibble: 2 × 2
#>          LD1        LD2
#>        <dbl>      <dbl>
#> 1 -1.1238094 -0.3321454
#> 2 -0.2653032 -1.1058991
#>